+++ src/report-builder.js
@@ -332,10 +332,21 @@ const createDocumentation = (ast, options) => {
       );
       return ebnfTemplate({
         identifier: production.identifier,
-        ebnf: productionToEBNF(production, {
-          markup: true,
-          format: options.textFormatting
-        }),
+        ebnf: [
+          productionToEBNF(
+            options.optimizeDiagrams
+              ? optimizeProduction(production, { textMode: true })
+              : production,
+            {
+              markup: true,
+              format: options.textFormatting
+            }
+          ),
+          productionToEBNF(production, {
+            markup: true,
+            format: options.textFormatting
+          })
+        ].join("\n"),
         referencedBy: searchReferencesToIdentifier(production.identifier, ast),
         referencesTo: outgoingReferences,
         diagram: vacuum(diagram.toString())