+++ src/report-builder.js
@@ -13,7 +13,6 @@ const {
   Stack,
   Terminal
 } = require("railroad-diagrams");
-
 const { optimizeProduction } = require("./structure-optimizer");
 const {
   documentContent,
@@ -332,10 +331,15 @@ const createDocumentation = (ast, options) => {
       );
       return ebnfTemplate({
         identifier: production.identifier,
-        ebnf: productionToEBNF(production, {
-          markup: true,
-          format: options.textFormatting
-        }),
+        ebnf: productionToEBNF(
+          options.optimizeText
+            ? optimizeProduction(production, { textMode: true })
+            : production,
+          {
+            markup: true,
+            format: options.textFormatting
+          }
+        ),
         referencedBy: searchReferencesToIdentifier(production.identifier, ast),
         referencesTo: outgoingReferences,
         diagram: vacuum(diagram.toString())