+++ CHANGELOG.md
@@ -12,6 +12,17 @@ and this project adheres to
 
 - Skip only diagram wrapping with `--no-diagram-wrap`
 - Breaking of long elements over multiple lines in optional items `[]`
+- Plain text will now also be optimized when reasonable:
+
+  ```
+  [ [ a ] ] => [ a ]
+  [ { a } ] => { a }
+  a | b | [ c | [ d | e ] ] => [ a | b | c | d | e ]
+  ```
+
+## Fixes
+
+- When elements in choices are grouped, it will now maintain the original order
 
 ### Changed