+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.11.1] - 2020-09-23
+
+## Fixes
+
+- Overview diagrams should also optimize its sub elements
+
 ## [1.11.0] - 2020-09-23
 
 ### Added